+++ CHANGELOG.md
@@ -11,6 +11,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - `MissingOsmCacheWarning` for the user to let him know that OSM index has to be built and the source can be switched to Geofabrik [#213](https://github.com/kraina-ai/quackosm/issues/213)
 
+### Changed
+
+- Bumped minimal version of the `h3` library to `4.1.0`
+
 ## [0.14.1] - 2025-05-23
 
 ### Added