+++ CHANGELOG.md
@@ -12,6 +12,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - `MissingOsmCacheWarning` for the user to inform that OSM index has to be built and the source can be switched to Geofabrik [#213](https://github.com/kraina-ai/quackosm/issues/213)
 - `OldOsmCacheWarning` for the user to inform that currently saved cache is over 1 year old and it can be outdated [#213](https://github.com/kraina-ai/quackosm/issues/213)
 
+### Changed
+
+- Bumped minimal version of the `h3` library to `4.1.0`
+
 ## [0.14.1] - 2025-05-23
 
 ### Added